function_extract/requirements.txt
azure-functions==1.11.0
azure-storage-blob==12.24.0
requests>=2.28.0
pandas>=2.0.0
